fix float coords issue

+++ checkers/Game.py
@@ -48,7 +48,7 @@ class Game:
 						self.board.move_piece(self.selected_piece, self.mouse_pos)
 					
 						if self.mouse_pos not in self.board.adjacent(self.selected_piece):
-							self.board.remove_piece((self.selected_piece[0] + (self.mouse_pos[0] - self.selected_piece[0]) / 2, self.selected_piece[1] + (self.mouse_pos[1] - self.selected_piece[1]) / 2))
+							self.board.remove_piece((int(self.selected_piece[0] + (self.mouse_pos[0] - self.selected_piece[0]) / 2), int(self.selected_piece[1] + (self.mouse_pos[1] - self.selected_piece[1]) / 2)))
 						
 							self.hop = True
 							self.selected_piece = self.mouse_pos
@@ -59,7 +59,7 @@ class Game:
 				if self.hop == True:					
 					if self.selected_piece != None and self.mouse_pos in self.board.legal_moves(self.selected_piece, self.hop):
 						self.board.move_piece(self.selected_piece, self.mouse_pos)
-						self.board.remove_piece((self.selected_piece[0] + (self.mouse_pos[0] - self.selected_piece[0]) / 2, self.selected_piece[1] + (self.mouse_pos[1] - self.selected_piece[1]) / 2))
+						self.board.remove_piece((int(self.selected_piece[0] + (self.mouse_pos[0] - self.selected_piece[0]) / 2), int(self.selected_piece[1] + (self.mouse_pos[1] - self.selected_piece[1]) / 2)))
 
 					if self.board.legal_moves(self.mouse_pos, self.hop) == []:
 							self.end_turn()
